Move comit-rs setting file up as it is now unique
