+++ src/cnd_settings.rs
