--- src/comit_rs_settings/mod.rs
@@ -1,1 +1,0 @@
-pub mod cnd;
