+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod comit_rs_settings;
+pub mod cnd_settings;
 pub mod create_comit_app;
 pub mod docker;
 pub mod new;
